Cambio algunos comentarios y corrijo la recepción del arma por socket

+++ server/game/game_logic.cpp
@@ -6,7 +6,7 @@
 void GameLogic::buy_weapon(FullPlayer& player, Model::WeaponID weapon_id, const Round& round) const {
     if (!round.is_buying()) return;
     if (!player.is_alive()) return;
-    // FALTA: Cuando este mapa ver si el jugador está en zona de compra
+    // TODO: Check player is in shop zone
 
     shop.process_weapon_purchase(player, weapon_id);
 }
@@ -27,7 +27,6 @@ void GameLogic::process_shooting(std::map<uint8_t, FullPlayer>& players, Round& 
 
         auto shot_info = player.shoot(frames_to_process);
         if (!shot_info.has_value()) continue;
-        // OJO FUEGO AMIGO PERMITIDO, NO SE SI SACAR ESO EN SHOTMANAGER O EN APPLY IMPACTS.
         std::vector<Impact> impacts = ShotManager::calculate_shot_impacts(shot_info.value(), players);
         if (!impacts.empty()) {
             apply_impacts(impacts, round, players);
@@ -45,7 +44,7 @@ void GameLogic::apply_impacts(const std::vector<Impact>& impacts, Round& round, 
 
         auto victim = players.find(impact.victim_id);
         if (victim == players.end()) continue;
-        //ACÁ DEBRÍA CHEQUEAR QUE NO SE FUEGO AMIGO, SI ES FUEGO AMIGO CONTINUE
+        if (victim->second.get_team() == shooter->second.get_team()) continue;
         if(!victim->second.is_alive()) continue; // no podes matar a un muerto, si no te sumaria una kill que no es tuya
 
         victim->second.take_damage(impact.damage);
